+++ app/Http/Controllers/Api/V1/LessonController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\Api\V1;
 
 use App\Http\Controllers\Controller;
-use App\Services\LessonServiceInterface;
+use App\Services\Interfaces\LessonServiceInterface;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use App\Exceptions\DataNotFoundException;
