+++ app/Http/Controllers/Api/V1/UserProgressController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\Api\V1;
 
 
 use App\Http\Controllers\Controller;
-use App\Services\UserProgressServiceInterface;
+use App\Services\Interfaces\UserProgressServiceInterface;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
