+++ app/Repositories/Interfaces/FriendRequestRepositoryInterface.php
@@ -1,6 +1,7 @@
 <?php
 
-namespace App\Repositories;
+namespace App\Repositories\Interfaces;
+
 
 use App\Models\FriendRequest;
