+++ app/Repositories/Interfaces/LessonRepositoryInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Repositories;
+namespace App\Repositories\Interfaces;
 
 use App\Models\Lesson;
 use Illuminate\Support\Collection;
