+++ app/Repositories/Interfaces/OptionRepositoryInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Repositories;
+namespace App\Repositories\Interfaces;
 
 use App\Models\Option;
 use Illuminate\Support\Collection;
