+++ app/Repositories/Interfaces/QuestionRepositoryInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Repositories;
+namespace App\Repositories\Interfaces;
 
 use App\Models\Question;
 use Illuminate\Support\Collection;
