+++ app/Repositories/Interfaces/UserProgressRepositoryInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Repositories;
+namespace App\Repositories\Interfaces;
 
 use App\Models\UserProgress;
 use Illuminate\Support\Collection;
