+++ app/Repositories/OptionRepository.php
@@ -3,6 +3,7 @@
 namespace App\Repositories;
 
 use App\Models\Option;
+use App\Repositories\Interfaces\OptionRepositoryInterface;
 use Illuminate\Support\Collection;
 
 class OptionRepository extends BaseRepository implements OptionRepositoryInterface
