+++ app/Services/UserProgressService.php
@@ -6,8 +6,9 @@ use App\DTOs\UserStatsDTO;
 use App\Exceptions\DataNotFoundException;
 use App\Exceptions\InvalidParamException;
 use App\Models\UserProgress;
-use App\Repositories\QuestionRepositoryInterface;
-use App\Repositories\UserProgressRepositoryInterface;
+use App\Repositories\Interfaces\QuestionRepositoryInterface;
+use App\Repositories\Interfaces\UserProgressRepositoryInterface;
+use App\Services\Interfaces\UserProgressServiceInterface;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
